Refactor mythril/mythril.py
