+++ mythril/mythril/__init__.py
@@ -1,0 +1,4 @@
+from .mythril_disassembler import MythrilDisassembler
+from .mythril_analyzer import MythrilAnalyzer
+from .mythril_config import MythrilConfig
+from .mythril_leveldb import MythrilLevelDB
