+++ tests/laser/transaction/create_transaction_test.py
@@ -1,4 +1,4 @@
-from mythril.mythril import Mythril
+from mythril.mythril import MythrilDisassembler
 from mythril.laser.ethereum.transaction import execute_contract_creation
 from mythril.ethereum import util
 import mythril.laser.ethereum.svm as svm
@@ -13,7 +13,7 @@ from mythril.analysis.symbolic import SymExecWrapper
 def test_create():
     contract = SolidityContract(
         str(tests.TESTDATA_INPUTS_CONTRACTS / "calls.sol"),
-        solc_binary=Mythril._init_solc_binary("0.5.0"),
+        solc_binary=MythrilDisassembler._init_solc_binary("0.5.0"),
     )
 
     laser_evm = svm.LaserEVM({})
@@ -37,7 +37,7 @@ def test_create():
 def test_sym_exec():
     contract = SolidityContract(
         str(tests.TESTDATA_INPUTS_CONTRACTS / "calls.sol"),
-        solc_binary=Mythril._init_solc_binary("0.5.0"),
+        solc_binary=MythrilDisassembler._init_solc_binary("0.5.0"),
     )
 
     sym = SymExecWrapper(
